refactor: 상세페이지 화면 섹션별로 분리하여 파일 작성 및 중복된 폴더명 변경 #64
