+++ gomintapa/lib/src/screens/auth/register.dart
@@ -2,10 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:gomintapa/src/screens/auth/login.dart';
 import 'package:gomintapa/src/widgets/themes/auth_theme.dart';
-import 'package:gomintapa/src/widgets/forms/label_textfield.dart';
+import 'package:gomintapa/src/widgets/forms/input/label_textfield.dart';
 
 import '../../controllers/auth_controller.dart';
-import '../../home.dart';
 
 class Register extends StatefulWidget {
   const Register({super.key});
